+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.3
 
 - Reference tuning changes no longer change the nominals mode my default. Added `!<midinote>: <freq>` syntax to change mode of the nominals.
+- Fixed nominals not showing up in enharmonic cycle when no important ligatures are in the enharmonic group.
 
 ## 0.2.2